pyproject.toml
[tool.poetry]
name = "hmw5"
version = "0.1.0"
description = ""
authors = ["Maria Scofa <mariascofa@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
Django = "3.1.4"
flake8 = "3.8.4"
Faker = "5.3.0"
gender-guesser = "0.4.0"
django-crispy-forms = "1.10.0"
celery = "5.0.5"
SQLAlchemy = "^1.3.22"

[tool.poetry.dev-dependencies]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
